refactor(Initializer): refactor initializer
