+++ python/oneflow/nn/init.py
@@ -16,21 +16,54 @@ limitations under the License.
 import os
 
 import oneflow as flow
-from oneflow.ops.initializer_util import CalcGain
+from oneflow.framework.tensor import _copy_from_numpy_to_eager_local_tensor, Tensor
+from oneflow.ops.util.initializer_util import calc_gain as calculate_gain
+import oneflow.ops.initializer_register as initializer_register
 
 
-def calculate_gain(nonlinearity, param=None):
-    return CalcGain(nonlinearity, param)
+def _init_by_initializer_conf(tensor, initializer_conf, random_seed=None):
+    # NOTE: initializing weight should not enable autograd mode
+    if random_seed is None:
+        random_seed = flow.default_generator.initial_seed()
+    shape = tuple(tensor.shape)
+    initializer = initializer_register.get_initializer(
+        initializer_conf, random_seed, shape
+    )
+
+    np_arr = initializer_register.generate_values_by_initializer(
+        initializer, shape, tensor.dtype
+    )
+    with flow.no_grad():
+        if tensor.is_global:
+            src_tensor = flow.tensor(np_arr)
+            src_tensor = src_tensor.to_global(
+                placement=tensor.placement,
+                sbp=tuple(flow.sbp.broadcast for _ in range(len(tensor.sbp))),
+            )
+            tensor.copy_(src_tensor)
+        else:
+            _copy_from_numpy_to_eager_local_tensor(
+                tensor, np_arr,
+            )
+    return tensor
 
 
 def uniform_(tensor, a=0.0, b=1.0):
-    with flow.no_grad():
-        return tensor.uniform_(a, b)
+    if isinstance(a, Tensor):
+        assert a.ndim == 0 and a.nelement() == 1, "a must be a number or scalar tensor!"
+        a = a.numpy().item()
+    if isinstance(b, Tensor):
+        assert b.ndim == 0 and b.nelement() == 1, "b must be a number or scalar tensor!"
+        b = b.numpy().item()
+    initializer_conf = initializer_register.random_uniform_initializer(
+        minval=a, maxval=b, dtype=tensor.dtype
+    )
+    return _init_by_initializer_conf(tensor, initializer_conf)
 
 
 def normal_(tensor, mean=0.0, std=1.0):
-    with flow.no_grad():
-        return tensor.normal_(mean, std)
+    initializer_conf = initializer_register.random_normal_initializer(mean, std)
+    return _init_by_initializer_conf(tensor, initializer_conf)
 
 
 def xavier_uniform_(tensor, gain=1.0, *, data_format="NCHW"):
@@ -57,8 +90,10 @@ def xavier_uniform_(tensor, gain=1.0, *, data_format="NCHW"):
         >>> w = flow.empty(3, 5)
         >>> nn.init.xavier_uniform_(w, gain=nn.init.calculate_gain('relu'))
     """
-    with flow.no_grad():
-        return tensor.xavier_uniform_(gain, data_format=data_format)
+    initializer_conf = initializer_register.xavier_initializer(tensor.shape, gain=gain,
+                                                               data_format=data_format,
+                                                               distribution="random_uniform")
+    return _init_by_initializer_conf(tensor, initializer_conf)
 
 
 def xavier_normal_(tensor, gain=1.0, *, data_format="NCHW"):
@@ -85,8 +120,10 @@ def xavier_normal_(tensor, gain=1.0, *, data_format="NCHW"):
         >>> w = flow.empty(3, 5)
         >>> nn.init.xavier_normal_(w)
     """
-    with flow.no_grad():
-        return tensor.xavier_normal_(gain, data_format=data_format)
+    initializer_conf = initializer_register.xavier_initializer(tensor.shape, gain=gain,
+                                                               data_format=data_format,
+                                                               distribution="random_normal")
+    return _init_by_initializer_conf(tensor, initializer_conf)
 
 
 def orthogonal_(tensor, gain=1.0):
@@ -145,8 +182,17 @@ def kaiming_uniform_(
         >>> w = flow.empty(3, 5)
         >>> nn.init.kaiming_uniform_(w, mode='fan_in', nonlinearity='relu')
     """
-    with flow.no_grad():
-        return tensor.kaiming_uniform_(a, mode, nonlinearity, data_format=data_format)
+    if os.getenv("ONEFLOW_ENABLE_NHWC") == "1":
+        data_format = "NHWC"
+    initializer_conf = initializer_register.kaiming_initializer(
+        tensor.shape,
+        a=a,
+        mode=mode,
+        nonlinearity=nonlinearity,
+        data_format=data_format,
+        distribution="random_uniform",
+    )
+    return _init_by_initializer_conf(tensor, initializer_conf)
 
 
 def kaiming_normal_(
@@ -184,13 +230,25 @@ def kaiming_normal_(
     """
     if os.getenv("ONEFLOW_ENABLE_NHWC") == "1":
         data_format = "NHWC"
-    with flow.no_grad():
-        return tensor.kaiming_normal_(a, mode, nonlinearity, data_format=data_format)
+    initializer_conf = initializer_register.kaiming_initializer(
+        tensor.shape,
+        a=a,
+        mode=mode,
+        nonlinearity=nonlinearity,
+        data_format=data_format,
+        distribution="random_normal",
+    )
+    return _init_by_initializer_conf(tensor, initializer_conf)
 
 
 def trunc_normal_(tensor, mean=0.0, std=1.0, a=-2.0, b=2.0):
-    with flow.no_grad():
-        return tensor.trunc_normal_(mean, std, a, b)
+    initializer_conf = initializer_register.truncated_normal_initializer(
+        mean=mean,
+        std=std,
+        a=a,
+        b=b,
+    )
+    return _init_by_initializer_conf(tensor, initializer_conf)
 
 
 def constant_(tensor, val):
