+++ python/oneflow/test/tensor/test_tensor_part_1.py
@@ -168,8 +168,10 @@ class TestTensor(flow.unittest.TestCase):
         flow.nn.init.kaiming_normal_(z, a=0.1, mode="fan_out", nonlinearity="relu")
         flow.nn.init.kaiming_uniform_(z)
         z.requires_grad_()
-        flow.nn.init.xavier_normal_(z)
-        flow.nn.init.xavier_uniform_(z)
+        flow.nn.init.xavier_normal_(z, flow.nn.init.calculate_gain('relu'))
+        flow.nn.init.xavier_uniform_(z, flow.nn.init.calculate_gain('relu'))
+        flow.nn.init.xavier_normal_(z, flow.nn.init.calculate_gain('leaky_relu', 0.2))
+        flow.nn.init.xavier_uniform_(z, flow.nn.init.calculate_gain('leaky_relu', 0.2))
         flow.nn.init.orthogonal_(z)
         x = tensor_creator(*shape).to(dtype=flow.int32)
         np_ones = np.ones(x.shape, dtype=np.int32)
@@ -1155,27 +1157,27 @@ class TestTensor(flow.unittest.TestCase):
     def test_tensor_constructor(test_case):
         x = flow.tensor([1, 2, 3])
         test_case.assertTrue(np.array_equal(x.numpy(), [1, 2, 3]))
-        test_case.assertEquals(x.dtype, flow.int64)
+        test_case.assertEqual(x.dtype, flow.int64)
         x = flow.tensor([1.0, 2.0, 3.0])
         test_case.assertTrue(np.array_equal(x.numpy(), [1.0, 2.0, 3.0]))
-        test_case.assertEquals(x.dtype, flow.float32)
+        test_case.assertEqual(x.dtype, flow.float32)
         x = flow.tensor([1.0, 2.0, 3.0], dtype=flow.float64)
         test_case.assertTrue(np.array_equal(x.numpy(), [1.0, 2.0, 3.0]))
-        test_case.assertEquals(x.dtype, flow.float64)
+        test_case.assertEqual(x.dtype, flow.float64)
         np_arr = np.array([1, 2, 3])
         x = flow.tensor(np_arr)
         test_case.assertTrue(np.array_equal(x.numpy(), [1, 2, 3]))
-        test_case.assertEquals(x.dtype, flow.int64)
+        test_case.assertEqual(x.dtype, flow.int64)
         np_arr = np.array([1, 2, 3], dtype=np.float64)
         x = flow.tensor(np_arr)
         test_case.assertTrue(np.array_equal(x.numpy(), [1.0, 2.0, 3.0]))
-        test_case.assertEquals(x.dtype, flow.float64)
+        test_case.assertEqual(x.dtype, flow.float64)
         x = flow.tensor(np_arr, dtype=flow.float32)
         test_case.assertTrue(np.array_equal(x.numpy(), [1.0, 2.0, 3.0]))
-        test_case.assertEquals(x.dtype, flow.float32)
+        test_case.assertEqual(x.dtype, flow.float32)
         x = flow.tensor(np_arr, dtype=flow.int8)
         test_case.assertTrue(np.array_equal(x.numpy(), [1.0, 2.0, 3.0]))
-        test_case.assertEquals(x.dtype, flow.int8)
+        test_case.assertEqual(x.dtype, flow.int8)
 
     @profile(torch.Tensor.fill_)
     def profile_fill_(test_case):
